+++ packages/web3-validator/src/constants.ts
@@ -7,3 +7,9 @@ export const VALID_ETH_BASE_TYPES = [
 	'address',
 	'tuple',
 ] as const;
+
+export enum BlockTags {
+	EARLIEST = 'earliest',
+	LATEST = 'latest',
+	PENDING = 'pending',
+}