+++ packages/web3-validator/src/formats/index.ts
@@ -1,1 +1,2 @@
 export * from './hex';
+export * from './number';