+++ packages/web3-validator/src/formats/number.ts
@@ -1,0 +1,4 @@
+import { FormatDefinition } from 'ajv';
+import { isUInt } from '../validation/numbers';
+
+export const uint: FormatDefinition<number> = { validate: (data: number) => isUInt(data) };